backend/tech-agent/requirements.txt
langgraph
langchain
langchain-openai
langchain-community
fastapi
uvicorn
pydantic
python-dotenv
pygithub
tavily-python
pytest
streamlit
requests
redis
qdrant-client[fastembed]>=1.9.0
tiktoken>=0.7.0
